backend/requirements.txt
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
et_xmlfile==2.0.0
Flask==3.1.1
flask-cors==6.0.1
Flask-Executor==1.0.0
Flask-SQLAlchemy==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.6
openpyxl==3.1.5
pandas==2.3.1
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
six==1.17.0
soupsieve==2.7
SQLAlchemy==2.0.41
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
xlrd==2.0.2
python-dotenv==1.1.1
